Refactor ProductDetail components: move Dropdown and Dropdownbutt to new directory structure and handle error for quantiy
